Move InlierThresholdOptions into the corresponding file and remove unnecessary includes.
