+++ src/glomap/estimators/bundle_adjustment.h
@@ -6,7 +6,6 @@
 #include "glomap/estimators/optimization_base.h"
 #include "glomap/scene/frame.h"
 #include "glomap/scene/types.h"
-#include "glomap/types.h"
 
 #include <ceres/ceres.h>
