+++ src/glomap/estimators/global_positioning.h
@@ -6,7 +6,6 @@
 #include "glomap/estimators/optimization_base.h"
 #include "glomap/scene/frame.h"
 #include "glomap/scene/view_graph.h"
-#include "glomap/types.h"
 
 namespace glomap {
