+++ src/glomap/estimators/global_rotation_averaging.h
@@ -5,7 +5,6 @@
 
 #include "glomap/scene/frame.h"
 #include "glomap/scene/view_graph.h"
-#include "glomap/types.h"
 
 #include <string>
 #include <vector>
