+++ src/glomap/estimators/gravity_refinement.h
@@ -5,7 +5,6 @@
 #include "glomap/estimators/optimization_base.h"
 #include "glomap/scene/frame.h"
 #include "glomap/scene/view_graph.h"
-#include "glomap/types.h"
 
 #include <ceres/ceres.h>
