+++ src/glomap/sfm/global_mapper.h
@@ -7,6 +7,7 @@
 #include "glomap/estimators/global_rotation_averaging.h"
 #include "glomap/estimators/relpose_estimation.h"
 #include "glomap/estimators/view_graph_calibration.h"
+#include "glomap/processors/image_pair_inliers.h"
 #include "glomap/sfm/track_establishment.h"
 #include "glomap/sfm/track_retriangulation.h"
